requirements.txt
python-telegram-bot==20.8
aiosqlite==0.20.0
python-dotenv==1.0.1
